% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildingFunctions.R
\name{createBuilding}
\alias{createBuilding}
\title{Create a School Building}
\usage{
createBuilding(schoolId, object = createBuildingObject())
}
\arguments{
\item{schoolId}{The ID of the school to which the building will be added.}

\item{object}{Must be created via createBuildingObject().}
}
\value{
A dataframe of details for the newly created building.
}
\description{
This function creates a new school building.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/building}{API Documentation}
}

