% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildingFunctions.R
\name{createBuildingObject}
\alias{createBuildingObject}
\title{Create Building Object}
\usage{
createBuildingObject(title = NULL, address1 = NULL, address2 = NULL,
  city = NULL, state = NULL, postal_code = NULL, country = NULL,
  website = NULL, phone = NULL, fax = NULL, picture_url = NULL)
}
\arguments{
\item{title, address1, address2, city, state, postal_code, country, website, phone, fax, picture_url}{See \href{https://developers.schoology.com/api-documentation/rest-api-v1/building}{API Documentation} for a description 
of each parameter.}
}
\value{
A named list of building attributes.
}
\description{
This function creates a Schoology building object.
}
\details{
A building object is necessary for creation and modification of a Schoology building.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/building}{API Documentation}
}

\concept{Buildings}
