% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrollmentFunctions.R
\name{createEnrollment}
\alias{createEnrollment}
\title{Create Enrollment}
\usage{
createEnrollment(realm, realm_id, uid, admin, status)
}
\arguments{
\item{realm}{One of 'schools', 'buildings', 'users', 'groups', 'courses', or 'sections'.}

\item{realm_id}{The id of the realm object.}

\item{uid}{The Schoology ID of the user.}

\item{admin}{Whether this user is an administrator (e.g. teacher) as opposed to a regular member (e.g. student)}

\item{status}{The current status of the enrollment. 1: Active, 2: Expired (i.e. past course), 3: Invite pending, 4: Request Pending, 5: Archived (Course specific status members can be placed in before being fully unenrolled)}
}
\value{
A dataframe of updated enrollment details.
}
\description{
This function modifies one or more attributes of a enrollment.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/course-enrollment}{API Documentation}
}

\concept{Enrollments}
