% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradingPeriodFunctions.R
\name{createGradingPeriod}
\alias{createGradingPeriod}
\title{Create a School GradingPeriods}
\usage{
createGradingPeriod(title, start, end)
}
\arguments{
\item{title}{The grading period title. This title must be unique across all grading periods in the school.}

\item{start}{The grading period start date YYYY-MM-DD}

\item{end}{The grading period end date YYYY-MM-DD}
}
\value{
A dataframe of details for the newly created gradingPeriod.
}
\description{
This function creates a new school gradingPeriods.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/gradingPeriods}{API Documentation}
}

\concept{GradingPeriods}
