% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrivateMessagingFunctions.R
\name{createMessage}
\alias{createMessage}
\title{Create a School Message}
\usage{
createMessage(schoolId, title = NULL, address1 = NULL,
  address2 = NULL, city = NULL, state = NULL, postal_code = NULL,
  country = NULL, website = NULL, phone = NULL, fax = NULL,
  picture_url = NULL)
}
\arguments{
\item{schoolId}{The ID of the school to which the message will be added.}

\item{title, address1, address2, city, state, postal_code, country, website, phone, fax, picture_url}{See \href{https://developers.schoology.com/api-documentation/rest-api-v1/message}{API Documentation} for a description 
of each parameter.}
}
\value{
A dataframe of details for the newly created message.
}
\description{
This function creates a new school message.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/message}{API Documentation}
}

\concept{Messages}
