% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SectionFunctions.R
\name{createSectionObject}
\alias{createSectionObject}
\title{Create Section Object}
\usage{
createSectionObject(id = NULL, section_title = NULL,
  section_code = NULL, section_school_code = NULL,
  access_code = NULL, grading_periods = NULL, description = NULL,
  profile_url = NULL, location = NULL, meeting_days = NULL,
  start_time = NULL, end_time = NULL, class_periods = NULL,
  synced = NULL)
}
\arguments{
\item{id, school_id, building_id, school_uid, name_title, name_title_show, name_first, name_first_preferred, name_middle, name_middle_show, name_last, name_display, sectionname, primary_email, position, gender, grad_year, birthday_date, password, role_id, email_login_info, profiel_url, tz_name, parents, parent_uids, advisor_uids, child_uids, send_message, synced, profile_picture_fid, additional_buildings}{See \href{https://developers.schoology.com/api-documentation/rest-api-v1/course-section}{API Documentation} for a description 
of each parameter.}
}
\value{
A named list of section attributes.
}
\description{
This function creates a Schoology section object.
}
\details{
A section object is necessary for creation and modification of a Schoology section.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/course-section}{API Documentation}
}

\concept{Sections}
