% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{createUser}
\alias{createUser}
\title{Create a New User}
\usage{
createUser(object = createUserObject())
}
\arguments{
\item{object}{Must be created via createUserObject().}
}
\value{
A dataframe of details for the newly created user.
}
\description{
This function creates a new Schoology user.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/user}{API Documentation}
}

