% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{createUserObject}
\alias{createUserObject}
\title{Create User Object}
\usage{
createUserObject(id = NULL, school_id = NULL, building_id = NULL,
  school_uid = NULL, name_title = NULL, name_title_show = NULL,
  name_first = NULL, name_first_preferred = NULL, name_middle = NULL,
  name_middle_show = NULL, name_last = NULL, name_display = NULL,
  username = NULL, primary_email = NULL, position = NULL,
  gender = NULL, grad_year = NULL, birthday_date = NULL,
  password = NULL, role_id = NULL, email_login_info = NULL,
  profiel_url = NULL, tz_name = NULL, parents = NULL,
  parent_uids = NULL, advisor_uids = NULL, child_uids = NULL,
  send_message = NULL, synced = NULL, profile_picture_fid = NULL,
  additional_buildings = NULL)
}
\arguments{
\item{id, school_id, building_id, school_uid, name_title, name_title_show, name_first, name_first_preferred, name_middle, name_middle_show, name_last, name_display, username, primary_email, position, gender, grad_year, birthday_date, password, role_id, email_login_info, profiel_url, tz_name, parents, parent_uids, advisor_uids, child_uids, send_message, synced, profile_picture_fid, additional_buildings}{See \href{https://developers.schoology.com/api-documentation/rest-api-v1/user}{API Documentation} for a description 
of each parameter.}
}
\value{
A named list of user attributes.
}
\description{
This function creates a Schoology user object.
}
\details{
A user object is necessary for creation and modification of a Schoology user.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/user}{API Documentation}

NOTE: Documentation is incorrect. id NOT school_uid is required for bulk updating users.
}

