% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{createUsers}
\alias{createUsers}
\title{Create Multiple Users}
\usage{
createUsers(userObjects = list(createUserObject()),
  update_existing = F, ignore_email_conflicts = F,
  email_conflict_resolution = T)
}
\arguments{
\item{userObjects}{A list of userObjects. These must be created via createUserObject().}

\item{update_existing}{}
}
\value{
A dataframe of updated user details for each user created.
}
\description{
This function creates one or more attributes of up to 50 users at once.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/user}{API Documentation}
}

\concept{Users}
