% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildingFunctions.R
\name{deleteBuilding}
\alias{deleteBuilding}
\title{Deletes a school building.}
\usage{
deleteBuilding(buildingId)
}
\arguments{
\item{buildingId}{The ID of the building which will be deleted.}
}
\value{
The success status of the DELETE request.
}
\description{
Deletes a school building.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/building}{API Documentation}
}

