% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradingPeriodFunctions.R
\name{deleteGradingPeriod}
\alias{deleteGradingPeriod}
\title{Deletes a school gradingPeriod.}
\usage{
deleteGradingPeriod(gradingPeriodId)
}
\arguments{
\item{gradingPeriodsId}{The ID of the gradingPeriods which will be deleted.}
}
\value{
The success status of the DELETE request.
}
\description{
Deletes a school gradingPeriod.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/gradingPeriods}{API Documentation}
}

\concept{GradingPeriods}
