% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrivateMessagingFunctions.R
\name{deleteMessage}
\alias{deleteMessage}
\title{Deletes a school message.}
\usage{
deleteMessage(messageId)
}
\arguments{
\item{messageId}{The ID of the message which will be deleted.}
}
\value{
The success status of the DELETE request.
}
\description{
Deletes a school message.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/message}{API Documentation}
}

\concept{Messages}
