% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthenticationFunctions.R
\name{deleteObject}
\alias{deleteObject}
\title{Delete Schoology Object}
\usage{
deleteObject(endpointWithQuery, consumerKey = myConsumerKey,
  consumerSecret = myConsumerSecret, token = "", tokenSecret = "")
}
\arguments{
\item{endpointWithQuery}{See \href{https://developers.schoology.com/api-documentation/rest-api-v1/}{API documentation}
for a list of endpoints.}
}
\value{
The DELETE request response.
}
\description{
This function deletes a Schoology object using a DELETE request.
}
\details{
This function is called by any Schoology SDK function requiring a DELETE request.
You will, in general, not need to call this function directly.
}
\section{Additional Arguments}{
 These arguments must be set via the "options" function prior to use of this function. e.g. options(consumerKey = "12345")\cr\cr
\strong{consumerKey and consumerSecret}\cr\cr For 2-legged authentication (on your behalf), these values can
be found at School Management -> Integration.\cr\cr For 3-legged authentication (on behalf of someone else via
your application), these values can be found at (yourSchoologyDomain)/apps/publisher.\cr\cr
\strong{token and tokenSecret}\cr\cr For 2-legged authentication, these should be left blank.\cr\cr
For 3-legged authentication, these are stored values for the current user.\cr\cr
See \href{https://developers.schoology.com/api-documentation/authentication}{authentication documentation} for details.
}

\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/}{Schoology Rest API Documentation}
}

\concept{Requests}
