% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SectionFunctions.R
\name{deleteSection}
\alias{deleteSection}
\title{Deletes a school section.}
\usage{
deleteSection(sectionId)
}
\arguments{
\item{sectionId}{The ID of the section which will be deleted.}
}
\value{
The success status of the DELETE request.
}
\description{
Deletes a school section.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/section}{API Documentation}
}

\concept{Sections}
