% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{deleteUser}
\alias{deleteUser}
\title{Deletes a User.}
\usage{
deleteUser(
  userId,
  option_comment = "",
  option_keep_enrollments = 1,
  email_notification = 0
)
}
\arguments{
\item{userId}{The ID of the user which will be deleted.}
}
\value{
The success status of the DELETE request.
}
\description{
Deletes a User.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/user}{API Documentation}
}

\concept{Users}
