% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{getObject}
\alias{getObject}
\title{Perform Schoology GET Request}
\usage{
getObject(endpointWithQuery, consumerKey, consumerSecret, token = "",
  tokenSecret = "")
}
\arguments{
\item{endpointWithQuery}{See \href{https://developers.schoology.com/api-documentation/rest-api-v1/}{API Documentation}
for a list of endpoints.}

\item{consumerKey, consumerSecret}{For 2-legged authentication (on your behalf), these values can
be found at School Management -> Integration.  For 3-legged authentication (on behalf of someone else via
your application), these values can be found at (yourSchoologyDomain)/apps/publisher.
See \href{https://developers.schoology.com/api-documentation/authentication}{Authentication Documentation} for details.}

\item{token, tokenSecret}{For 2-legged authentication, these should be left blank. 
For 3-legged authentication, these are stored values for the current user.
See \href{https://developers.schoology.com/api-documentation/authentication}{Authentication Documentation} for details.}
}
\value{
The content of the GET request response.
}
\description{
GET Request for a Schoology Endpoint
}
\details{
This function is called by any Schoology SDK function requiring a GET request.
You will, in general, not need to call this function directly.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/school}{API Documentation}
}

