% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttendanceFunctions.R
\name{listAttendances}
\alias{listAttendances}
\title{List Attendances for a Course}
\usage{
listAttendances(section_id, start = NULL, end = NULL, enrollment_id = NULL)
}
\arguments{
\item{section_id}{The id of the section for which the attendance entries are requested.}

\item{start}{Filter statuses for a given date range, inclusive YYYY-MM-DD.}

\item{end}{Filter statuses for a given date range, inclusive YYYY-MM-DD.}

\item{enrollment_id}{Filter statuses for a given enrollment.}
}
\value{
A dataframe of attendance details.
}
\description{
This function returns a list of attendances for a course past and present.
}
\details{
An attendance entry is a student record of being absent, late or excused late.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/course-attendance}{API Documentation}
}

\concept{Attendances}
