% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildingFunctions.R
\name{listBuildings}
\alias{listBuildings}
\title{List Buildings for School}
\usage{
listBuildings(schoolId, start = 0, limit = 200)
}
\arguments{
\item{start}{Index of first record to return. Min value is 0.}

\item{limit}{Number of records to return. 
Default is 20. Max value is 200.}
}
\value{
A dataframe of building details.
}
\description{
This function lists all buildings for a school district.
}
\details{
A building is like a campus of a school district.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/building}{API Documentation}
}

