% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrollmentFunctions.R
\name{listEnrollments}
\alias{listEnrollments}
\title{List Enrollments for a Course}
\usage{
listEnrollments(realm, realm_id, uid = NULL, enrollment_status = NULL,
  type = NULL, picture_size = NULL)
}
\arguments{
\item{realm}{One of 'schools', 'buildings', 'users', 'groups', 'courses', or 'sections'.}

\item{realm_id}{The id of the realm object.}

\item{uid}{Filter enrollments for a given user.}

\item{enrollment_status}{Filter enrollments for a given enrollment status 1-5.}

\item{type}{Filter enrollments by an enrollment type. Possible values: 'admin' or 'member'.}

\item{picture_size}{Specify size of profile picture returned with this user. Possible values: 'big', 'reg', 'sm', or 'tiny'.}
}
\value{
A dataframe of enrollment details.
}
\description{
This function returns a list of enrollments for a course past and present.
}
\details{
A enrollment is specific user assigned to the realm object.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/course-enrollment}{API Documentation}
}

\concept{Enrollments}
