% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grading.R
\name{listGrades}
\alias{listGrades}
\title{List Grades for Section}
\usage{
listGrades(
  section_id,
  assignment_id = NULL,
  enrollment_id = NULL,
  timestamp = NULL
)
}
\arguments{
\item{section_id}{The id of the course section.}

\item{assignment_id}{Filter grades for a given assignment}

\item{enrollment_id}{Filter grades for a given enrollment}

\item{timestamp}{Return only grades that have been changed since the given timestamp, according to the server time.}
}
\value{
A list of section grades.
}
\description{
Return all grades for all enrollments of a section.
}
\concept{Grading}
