% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradingPeriodFunctions.R
\name{listGradingPeriods}
\alias{listGradingPeriods}
\title{List GradingPeriods for School}
\usage{
listGradingPeriods(title = NULL, startswith = NULL)
}
\arguments{
\item{title}{Specifies an exact title to search for.}

\item{startswith}{Set this value to 1 to match all grading periods whose titles begin with the title provided.}
}
\value{
A dataframe of gradingPeriods details.
}
\description{
This function lists all gradingPeriods for a school district.
}
\details{
A gradingPeriod is like a semester or trimester for a given school year.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/gradingPeriods}{API Documentation}
}

\concept{GradingPeriods}
