% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grading.R
\name{listGradingScales}
\alias{listGradingScales}
\title{List Grading Scales}
\usage{
listGradingScales(section_id)
}
\arguments{
\item{section_id}{The id of the section.}
}
\value{
The grading scales for the section.
}
\description{
List grading scales for a section
}
\concept{Grading}
