% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupFunctions.R
\name{listGroups}
\alias{listGroups}
\title{List Groups for a School}
\usage{
listGroups(building_id = NULL)
}
\arguments{
\item{building_id}{Return only groups for the given building_id.}
}
\value{
A dataframe of group details.
}
\description{
This function returns a list of groups for a school past and present.
}
\details{
A group is a group meeting.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/groups}{API Documentation}
}

\concept{Groups}
