% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrivateMessagingFunctions.R
\name{listMessages}
\alias{listMessages}
\title{List Messages for School}
\usage{
listMessages(uid, messageFolder = "inbox")
}
\arguments{
\item{schoolId}{The ID of the school (district) for which to list messages.}
}
\value{
A dataframe of message details.
}
\description{
This function lists all messages for a school district.
}
\details{
A message is like a campus of a school district.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/message}{API Documentation}
}

\concept{Messages}
