% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoleFunctions.R
\name{listRoles}
\alias{listRoles}
\title{List Roles for District}
\usage{
listRoles()
}
\value{
A dataframe of role details.
}
\description{
This function lists all roles for a district.
}
\details{
A "role" is a thing in Schoology.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/role}{API Documentation}
}

\concept{Roles}
