% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchoolFunctions.R
\name{listSchools}
\alias{listSchools}
\title{List Schools for District}
\usage{
listSchools()
}
\value{
A dataframe of school details.
}
\description{
This function lists all schools for a district.
}
\details{
A "school" is a district in Schoology so this will simply
   return details about the (one) school district.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/school}{API Documentation}
}

\concept{Schools}
