% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SectionFunctions.R
\name{listSectionGrades}
\alias{listSectionGrades}
\title{List Section Grades}
\usage{
listSectionGrades(section_id, return_response = F)
}
\arguments{
\item{section_id}{The id of the course section.}
}
\value{
A list of section grades.
}
\description{
Return all grades for all enrollments of a section.
}
