% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttendanceFunctions.R
\name{updateAttendance}
\alias{updateAttendance}
\title{Update Attendance Details}
\usage{
updateAttendance(section_id, enrollment_id, date, status, comment = "")
}
\arguments{
\item{section_id}{The id of the section for which the attendance entry will be added.}

\item{enrollment_id}{The id of the enrollment for the student in the course section.}

\item{date}{The date of the attendance entry in YYYY-MM-DD format.}

\item{status}{1, 2, 3, or 4 corresponding to 'present', 'absent', 'late', or 'excused'.}
}
\value{
An array of attendance ids and statuses
}
\description{
This function modifies one or more attributes of a attendance.
}
\concept{Attendances}
