% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildingFunctions.R
\name{updateBuilding}
\alias{updateBuilding}
\title{Update Building Details}
\usage{
updateBuilding(buildingId, object = createBuildingObject())
}
\arguments{
\item{buildingId}{Can be found by navigating to the Schoology building information page.}

\item{object}{Must be created via createBuildingObject().}
}
\value{
A dataframe of updated building details.
}
\description{
This function modifies one or more attributes of a school building.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/building}{API Documentation}
}

