% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CourseFunctions.R
\name{updateCourse}
\alias{updateCourse}
\title{Update Course Details}
\usage{
updateCourse(
  courseId,
  course_title = NULL,
  course_code = NULL,
  course_school_code = NULL,
  access_code = NULL,
  grading_periods = NULL,
  description = NULL,
  profile_url = NULL,
  location = NULL,
  meeting_days = NULL,
  start_time = NULL,
  end_time = NULL,
  class_periods = NULL,
  synced = NULL
)
}
\arguments{
\item{courseId}{Can be found by navigating to the Schoology course information page.}

\item{course_title}{The course title (e.g. Course 10b).}

\item{course_code}{The course code must be unique across the course and grading period (e.g. "Spring 2010" can only have one "10b" for course "ENG101").}

\item{course_school_code}{The course school code must be unique across the school.}

\item{grading_periods}{The grading period IDs with which this course is associated.}

\item{description}{The course description.}

\item{location}{The location of the course course.}

\item{meeting_days}{Days of the week that this course course meets. Each day is represented by a number (Sun is 0, Sat is 7).}

\item{start_time}{The time this course starts on the specified meeting days, from 00:00 to 23:59 (local time). Not applicable if your school uses class period blocks (i.e. courses meet at different times on different days).}

\item{end_time}{The time this course ends on the specified meeting days.}

\item{class_periods}{The class period IDs with which this course is associated. Not applicable if your school does not use class period blocks (see start_time).}

\item{synced}{Whether or not this course course was synced with an external system (eg, Student Information System). The default value is 0. For synced course courses, the Course School Code field is not editable through Schoology.}
}
\value{
A dataframe of updated course details.
}
\description{
This function modifies one or more attributes of a course.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/courses}{API Documentation}
}

\concept{Courses}
