% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrollmentFunctions.R
\name{updateEnrollment}
\alias{updateEnrollment}
\title{Update Enrollment Details}
\usage{
updateEnrollment(enrollmentId, enrollment_title = NULL,
  enrollment_code = NULL, enrollment_school_code = NULL,
  access_code = NULL, grading_periods = NULL, description = NULL,
  profile_url = NULL, location = NULL, meeting_days = NULL,
  start_time = NULL, end_time = NULL, class_periods = NULL,
  synced = NULL)
}
\arguments{
\item{enrollmentId}{Can be found by navigating to the Schoology enrollment information page.}

\item{enrollment_title}{The enrollment title (e.g. Enrollment 10b).}

\item{enrollment_code}{The enrollment code must be unique across the course and grading period (e.g. "Spring 2010" can only have one "10b" for course "ENG101").}

\item{enrollment_school_code}{The enrollment school code must be unique across the school.}

\item{grading_periods}{The grading period IDs with which this course is associated.}

\item{description}{The enrollment description.}

\item{location}{The location of the course enrollment.}

\item{meeting_days}{Days of the week that this course enrollment meets. Each day is represented by a number (Sun is 0, Sat is 7).}

\item{start_time}{The time this enrollment starts on the specified meeting days, from 00:00 to 23:59 (local time). Not applicable if your school uses class period blocks (i.e. enrollments meet at different times on different days).}

\item{end_time}{The time this enrollment ends on the specified meeting days.}

\item{class_periods}{The class period IDs with which this course is associated. Not applicable if your school does not use class period blocks (see start_time).}

\item{synced}{Whether or not this course enrollment was synced with an external system (eg, Student Information System). The default value is 0. For synced course enrollments, the Enrollment School Code field is not editable through Schoology.}
}
\value{
A dataframe of updated enrollment details.
}
\description{
This function modifies one or more attributes of a enrollment.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/course-enrollment}{API Documentation}
}

\concept{Enrollments}
