% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SectionFunctions.R
\name{updateFinalComments}
\alias{updateFinalComments}
\title{Update Final Comments}
\usage{
updateFinalComments(
  section_id,
  period_id,
  enrollment_ids,
  comments,
  comment_statuses
)
}
\arguments{
\item{section_id}{The id of the course section.}

\item{grades}{The updated grades object using patch semantics.}
}
\value{
The updated final comments.
}
\description{
Set/modify comments for students for a grading period and the overall course.
}
