% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grading.R
\name{updateGrades}
\alias{updateGrades}
\title{Update Grades for Section}
\usage{
updateGrades(section_id, grades)
}
\arguments{
\item{section_id}{The id of the course section.}
}
\value{
The updated grades.
}
\description{
Update grades for a section.
}
\concept{Grading}
