% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradingPeriodFunctions.R
\name{updateGradingPeriods}
\alias{updateGradingPeriods}
\title{Update GradingPeriods Details}
\usage{
updateGradingPeriods(gradingPeriodId, title = NULL, start = NULL, end = NULL)
}
\arguments{
\item{title}{The grading period title. This title must be unique across all grading periods in the school.}

\item{start}{The grading period start date YYYY-MM-DD}

\item{end}{The grading period end date YYYY-MM-DD
See \href{https://developers.schoology.com/api-documentation/rest-api-v1/gradingPeriods}{API Documentation} for a description 
of each parameter.}

\item{gradingPeriodsId}{Can be found by navigating to the Schoology gradingPeriods information page.}
}
\value{
A dataframe of updated gradingPeriod details.
}
\description{
This function modifies one or more attributes of a school gradingPeriods.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/gradingPeriods}{API Documentation}
}

\concept{GradingPeriods}
