% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupFunctions.R
\name{updateGroup}
\alias{updateGroup}
\title{Update Group Details}
\usage{
updateGroup(groupId, building_id = NULL, school_id = NULL,
  title = NULL, picture_url = NULL, description = NULL,
  website = NULL, access_code = NULL, privacy_level = NULL,
  category = NULL, options_invite_type = NULL,
  options_member_post = NULL, options_member_post_comment = NULL,
  options_create_discussion = NULL, options_create_files = NULL)
}
\arguments{
\item{groupId}{Can be found by navigating to the Schoology group information page.}

\item{building_id}{The internal Schoology ID of the school building to which the group belongs.}

\item{school_id}{The internal Schoology ID of the school to which the group belongs.}

\item{title}{The title of the group.}

\item{picture_url}{The URL of the group's profile picture.}

\item{description}{The group description.}

\item{website}{The group website.}

\item{access_code}{The access code that users can use to join the group (only admins can see this value).}

\item{privacy_level}{The privacy of the group. One of 'everyone', 'school', 'building', or 'group'.}

\item{category}{The category of the group.}

\item{options_invite_type}{How members can join the group. 0: Invite only, 1: Request to join, 2: Anyone can join. Default 0.}

\item{options_member_post}{Whether or not a group member can post a group update. Default 1.}

\item{options_member_post_comment}{Whether or not a group member can post comments to group udpates. Default 1.}

\item{options_create_discussion}{Whether or not a group member can create a discussion thread. Default 0.}

\item{options_create_files}{Whether or not members can create resources for the group. Default 0.}
}
\value{
A dataframe of updated group details.
}
\description{
This function modifies one or more attributes of a group.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/groups}{API Documentation}
}

\concept{Groups}
