% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SectionFunctions.R
\name{updateSection}
\alias{updateSection}
\title{Update Section Details}
\usage{
updateSection(sectionId, object = createSectionObject())
}
\arguments{
\item{sectionId}{Can be found by navigating to the Schoology section information page.}

\item{object}{Must be created via createSectionObject().}
}
\value{
A dataframe of updated section details.
}
\description{
This function modifies one or more attributes of a section.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/course-section}{API Documentation}
}

\concept{Sections}
