% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{updateUser}
\alias{updateUser}
\title{Update User Details}
\usage{
updateUser(userId, object = createUserObject())
}
\arguments{
\item{userId}{Can be found by navigating to the Schoology user information page.}

\item{object}{Must be created via createUserObject().}
}
\value{
A dataframe of updated user details.
}
\description{
This function modifies one or more attributes of a user.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/user}{API Documentation}
}

\concept{Users}
