% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{updateUsers}
\alias{updateUsers}
\title{Update Multiple Users}
\usage{
updateUsers(userObjects = list(createUserObject()))
}
\arguments{
\item{userObjects}{A list of userObjects. These must be created via createUserObject().}
}
\value{
A dataframe of updated user details for each user modified.
}
\description{
This function modifies one or more attributes of up to 50 users at once.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/user}{API Documentation}
}

\concept{Users}
