% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildingFunctions.R
\name{viewBuilding}
\alias{viewBuilding}
\title{Get Building Details}
\usage{
viewBuilding(buildingId)
}
\arguments{
\item{buildingId}{Can be found by navigating to the Schoology building information page.}
}
\value{
A dataframe of building details.
}
\description{
This function returns details about a building: a campus of a school district.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/building}{API Documentation}
}

\concept{Buildings}
