% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CourseFunctions.R
\name{viewCourse}
\alias{viewCourse}
\title{Get Course Details}
\usage{
viewCourse(courseId)
}
\arguments{
\item{courseId}{Can be found by navigating to the Schoology course's information page.}
}
\value{
A dataframe of course details.
}
\description{
This function returns details about a Schoology course.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/courses}{API Documentation}
}

\concept{Courses}
