% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrollmentFunctions.R
\name{viewEnrollment}
\alias{viewEnrollment}
\title{Get Enrollment Details}
\usage{
viewEnrollment(realm, realm_id, enrollmentId)
}
\arguments{
\item{realm}{One of 'schools', 'buildings', 'users', 'groups', 'courses', or 'sections'.}

\item{realm_id}{The id of the realm object.}

\item{enrollmentId}{Can be found by navigating to the Schoology enrollment's information page.}
}
\value{
A dataframe of enrollment details.
}
\description{
This function returns details about a Schoology enrollment.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/course-enrollment}{API Documentation}
}

\concept{Enrollments}
