% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradingPeriodFunctions.R
\name{viewGradingPeriod}
\alias{viewGradingPeriod}
\title{Get GradingPeriods Details}
\usage{
viewGradingPeriod(gradingPeriodId)
}
\arguments{
\item{gradingPeriodsId}{Can be found by navigating to the Schoology gradingPeriods information page.}
}
\value{
A dataframe of gradingPeriods details.
}
\description{
This function returns details about a gradingPeriods: a campus of a school district.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/gradingPeriods}{API Documentation}
}

\concept{GradingPeriods}
