% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupFunctions.R
\name{viewGroup}
\alias{viewGroup}
\title{Get Group Details}
\usage{
viewGroup(groupId)
}
\arguments{
\item{groupId}{Can be found by navigating to the Schoology group's information page.}
}
\value{
A dataframe of group details.
}
\description{
This function returns details about a Schoology group.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/groups}{API Documentation}
}

\concept{Groups}
