% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrivateMessagingFunctions.R
\name{viewMessage}
\alias{viewMessage}
\title{Get Message Details}
\usage{
viewMessage(messageId, messageFolder)
}
\arguments{
\item{messageId}{Can be found by navigating to the Schoology message information page.}
}
\value{
A dataframe of message details.
}
\description{
This function returns details about a message: a campus of a school district.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/message}{API Documentation}
}

\concept{Messages}
