% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoleFunctions.R
\name{viewRole}
\alias{viewRole}
\title{Get Role Details}
\usage{
viewRole(roleId)
}
\arguments{
\item{roleId}{Can be found by navigating to the Roleogy role (district) information page.}
}
\value{
A dataframe of role (district) details.
}
\description{
This function returns details about a role (district).
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/role}{API Documentation}
}

\concept{Roles}
