% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchoolFunctions.R
\name{viewSchool}
\alias{viewSchool}
\title{Get School Details}
\usage{
viewSchool(schoolId)
}
\arguments{
\item{schoolId}{character. (SchoologyURL)/school/(schoolId)}
}
\value{
A dataframe of school details.
}
\description{
View details about a school (district).
}
\details{
(none)
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/school}{API Documentation}
}

