% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SectionFunctions.R
\name{viewSection}
\alias{viewSection}
\title{Get Section Details}
\usage{
viewSection(sectionId)
}
\arguments{
\item{sectionId}{Can be found by navigating to the Schoology section's information page.}
}
\value{
A dataframe of section details.
}
\description{
This function returns details about a Schoology section.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/course-section}{API Documentation}
}

\concept{Section}
\concept{Sections}
