% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{viewUser}
\alias{viewUser}
\title{Get User Details}
\usage{
viewUser(userId, active = T, extended = F)
}
\arguments{
\item{userId}{Can be found by navigating to the Schoology user's information page.}

\item{active}{Whether to return active users only. Defaults to true.}

\item{extended}{Whether to return extended details about a user. Defaults to false.}
}
\value{
A dataframe of user details.
}
\description{
This function returns details about a Schoology user.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/user}{API Documentation}
}

\concept{Users}
