% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildingFunctions.R
\name{listBuildings}
\alias{listBuildings}
\title{List Buildings for School}
\usage{
listBuildings(schoolId)
}
\arguments{
\item{schoolId}{The ID of the school (district) for which to list buildings.}
}
\value{
A dataframe of building details.
}
\description{
This function lists all buildings for a school district.
}
\details{
A building is like a campus of a school district.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/building}{API Documentation}
}

\concept{Buildings}
