% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchoolFunctions.R
\name{listSchools}
\alias{listSchools}
\title{List Schools for District}
\usage{
listSchools(start = 0, limit = 200)
}
\arguments{
\item{start}{Index of first record to return. Min value is 0.}

\item{limit}{Number of records to return. 
Default is 20. Max value is 200.}
}
\value{
A dataframe of school details.
}
\description{
This function lists all schools for a district.
}
\details{
A "school" is a district in Schoology so this will simply
   return details about the (one) school district.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/school}{API Documentation}
}

