% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SectionFunctions.R
\name{listSections}
\alias{listSections}
\title{List Sections for a Course}
\usage{
listSections(course_id, include_past = F)
}
\arguments{
\item{include_past}{Whether to return past sections or only current sections.}
}
\value{
A dataframe of section details.
}
\description{
This function returns a list of sections for a course past and present.
}
\details{
A section is a course meeting.
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/course-section}{API Documentation}
}

\concept{Sections}
