% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchoolFunctions.R
\name{updateSchool}
\alias{updateSchool}
\title{Update School Details}
\usage{
updateSchool(schoolId, object = createSchoolObject())
}
\arguments{
\item{object}{Must be created via createSchoolObject().}

\item{id}{Can be found by navigating to the Schoology district information page.}
}
\value{
A dataframe of updated school details.
}
\description{
This function modifies one or more attributes of a school (district).
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/school}{API Documentation}
}

