% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchoolFunctions.R
\name{updateSchool}
\alias{updateSchool}
\title{Update School Details}
\usage{
updateSchool(
  schoolId,
  title = NULL,
  address1 = NULL,
  address2 = NULL,
  city = NULL,
  state = NULL,
  postal_code = NULL,
  country = NULL,
  website = NULL,
  phone = NULL,
  fax = NULL,
  picture_url = NULL
)
}
\arguments{
\item{schoolId}{Can be found by navigating to the Schoology district information page.}

\item{title, address1, address2, city, state, postal_code, country, website, phone, fax, picture_url}{See \href{https://developers.schoology.com/api-documentation/rest-api-v1/school}{API Documentation} for a description 
of each parameter.}
}
\value{
A dataframe of updated school details.
}
\description{
This function modifies one or more attributes of a school (district).
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/school}{API Documentation}
}

\concept{Schools}
