% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchoolFunctions.R
\name{viewSchool}
\alias{viewSchool}
\title{Get School Details}
\usage{
viewSchool(schoolId)
}
\arguments{
\item{id}{Can be found by navigating to the Schoology school (district) information page.}
}
\value{
A dataframe of school (district) details.
}
\description{
This function returns details about a school (district).
}
\section{References}{

\href{https://developers.schoology.com/api-documentation/rest-api-v1/school}{API Documentation}
}

